Move timer functions from API to data flows
Jira ticket: CAMS-421
