+++ backend/function-apps/dataflows/import/office-staff-sync/office-staff-sync.function.test.ts
@@ -1,9 +1,9 @@
-import { LoggerImpl } from '../../../lib/adapters/services/logger.service';
-import { CamsError } from '../../../lib/common-errors/cams-error';
-import timerTrigger from './office-staff-sync.function';
 import { Timer } from '@azure/functions';
-import { createMockAzureFunctionContext } from '../../azure/testing-helpers';
-import { OfficesController } from '../../../lib/controllers/offices/offices.controller';
+import timerTrigger from './office-staff-sync.function';
+import { createMockAzureFunctionContext } from '../../../azure/testing-helpers';
+import { OfficesController } from '../../../../lib/controllers/offices/offices.controller';
+import { CamsError } from '../../../../lib/common-errors/cams-error';
+import { LoggerImpl } from '../../../../lib/adapters/services/logger.service';
 
 describe('Office Staff Sync Function tests', () => {
   const context = createMockAzureFunctionContext();
