+++ backend/function-apps/dataflows/import/office-staff-sync/office-staff-sync.function.ts
@@ -1,15 +1,15 @@
 import * as dotenv from 'dotenv';
 import { app, InvocationContext, Timer } from '@azure/functions';
-import ContextCreator from '../../azure/application-context-creator';
-import { initializeApplicationInsights } from '../../azure/app-insights';
-import { OrdersController } from '../../../lib/controllers/orders/orders.controller';
-import { toAzureError } from '../../azure/functions';
+import { initializeApplicationInsights } from '../../../azure/app-insights';
+import ContextCreator from '../../../azure/application-context-creator';
+import { OfficesController } from '../../../../lib/controllers/offices/offices.controller';
+import { toAzureError } from '../../../azure/functions';
 
 dotenv.config();
 
 initializeApplicationInsights();
 
-const MODULE_NAME = 'ORDERS-SYNC-FUNCTION';
+const MODULE_NAME = 'OFFICE-STAFF-SYNC-FUNCTION';
 
 export default async function timerTrigger(
   _myTimer: Timer,
@@ -18,14 +18,14 @@ export default async function timerTrigger(
   const logger = ContextCreator.getLogger(invocationContext);
   try {
     const appContext = await ContextCreator.getApplicationContext({ invocationContext, logger });
-    const ordersController = new OrdersController(appContext);
-    await ordersController.handleTimer(appContext);
+    const controller = new OfficesController();
+    await controller.handleTimer(appContext);
   } catch (error) {
     toAzureError(logger, MODULE_NAME, error);
   }
 }
 
-app.timer('orders-sync', {
-  schedule: '0 30 9 * * *',
+app.timer('office-staff-sync', {
+  schedule: '0 0 5 * * *',
   handler: timerTrigger,
 });
