+++ backend/function-apps/dataflows/import/orders-sync/orders-sync.function.ts
@@ -1,15 +1,15 @@
 import * as dotenv from 'dotenv';
 import { app, InvocationContext, Timer } from '@azure/functions';
-import ContextCreator from '../../azure/application-context-creator';
-import { initializeApplicationInsights } from '../../azure/app-insights';
-import { toAzureError } from '../../azure/functions';
-import { OfficesController } from '../../../lib/controllers/offices/offices.controller';
+import { initializeApplicationInsights } from '../../../azure/app-insights';
+import ContextCreator from '../../../azure/application-context-creator';
+import { OrdersController } from '../../../../lib/controllers/orders/orders.controller';
+import { toAzureError } from '../../../azure/functions';
 
 dotenv.config();
 
 initializeApplicationInsights();
 
-const MODULE_NAME = 'OFFICE-STAFF-SYNC-FUNCTION';
+const MODULE_NAME = 'ORDERS-SYNC-FUNCTION';
 
 export default async function timerTrigger(
   _myTimer: Timer,
@@ -18,14 +18,14 @@ export default async function timerTrigger(
   const logger = ContextCreator.getLogger(invocationContext);
   try {
     const appContext = await ContextCreator.getApplicationContext({ invocationContext, logger });
-    const controller = new OfficesController();
-    await controller.handleTimer(appContext);
+    const ordersController = new OrdersController(appContext);
+    await ordersController.handleTimer(appContext);
   } catch (error) {
     toAzureError(logger, MODULE_NAME, error);
   }
 }
 
-app.timer('office-staff-sync', {
-  schedule: '0 0 5 * * *',
+app.timer('orders-sync', {
+  schedule: '0 30 9 * * *',
   handler: timerTrigger,
 });
